Use Theme type instead of any
